refactor: organize adapter shared types
